fix failed test case
