+++ sdk/ml/azure-ai-ml/tests/dsl/e2etests/test_controlflow_pipeline.py
@@ -398,39 +398,6 @@ class TestIfElse(TestControlFlowPipeline):
             "type": "if_else",
         }
 
-    def test_if_else_with_group_input(self, client: MLClient):
-        hello_world_component_no_paths = load_component(
-            source=r"./tests/test_configs/components/helloworld_component_no_paths.yml"
-        )
-
-        @group
-        class SubGroup:
-            num: int
-
-        @group
-        class ParentGroup:
-            input_group: SubGroup
-
-        @pipeline(
-            compute="cpu-cluster",
-        )
-        def condition_pipeline(group_input: ParentGroup):
-            node1 = hello_world_component_no_paths(component_in_number=1)
-            condition(condition=group_input.input_group.num < 100, true_block=node1)
-
-        pipeline_job = condition_pipeline(group_input=ParentGroup(input_group=SubGroup(num=10)))
-        with include_private_preview_nodes_in_pipeline():
-            pipeline_job = assert_job_cancel(pipeline_job, client)
-
-        dsl_pipeline_job_dict = omit_with_wildcard(pipeline_job._to_rest_object().as_dict(), *omit_fields)
-        assert dsl_pipeline_job_dict["properties"]["jobs"]["expression_component"] == {
-            "environment_variables": {"AZURE_ML_CLI_PRIVATE_FEATURES_ENABLED": "true"},
-            "name": "expression_component",
-            "type": "command",
-            "inputs": {"num": {"job_input_type": "literal", "value": "${{parent.inputs.group_input.input_group.num}}"}},
-            "_source": "REMOTE.WORKSPACE.COMPONENT",
-        }
-
 
 @pytest.mark.usefixtures("mock_anon_component_version")
 class TestDoWhilePipeline(TestControlFlowPipeline):
